1. In main.c, corrected calls to random_gaussian_clocknoise() and random_uniform_clocknoise() to increase range inputs by a factor of 2. Noise modulation index was 50% of input value and is now accurate.
2. In main.c, added pre-processor commands for new DEBUG_SQUARE_WAVE #define
to only plot and save signal vout if DEBUG_SQUARE_WAVE is not defined.

3, In main.c, based on #define DEBUG_SQUARE_WAVE, changed  new gnuplot command
file 'gnu_plot.gnu" to only plot vout; renamed existing plot command file to "gnu_plot_debug.gnu"
and call that file if DEBUG_SQUARE_WAVE is defined.

4. In main.c, updated psd command lines to create its input file from column
5 of output file if DEBUG_SQUARE_WAVE is defined, or to use existing output
file containing only vout if DEBUG_SQUARE_WAVE is not defined.

5. In main.c, depending on definition of DEBUG_SQUARE_WAVE, remove psd input
file if a new one was created from column 5 of the output file. There is no
need to remove the file if the output file only contains time and vout.

6. In main.c, corrected arguments to check_inputs.c for new function prototype

7. In main.c, changed arguments and logic to allow function call with and without
noise arguments. The number of arguments if modulation index is 0 is reduced. Added
command line logic to accept variable number of arguments

8. In vpulse.h, added optional #define DEBUG_SQUARE_WAVE and commented out
definition

9. In vpulse.h, updated version and date to v1.7 adn 9/22/2023 and updated number
of decimals for version number to 3

10. Updated function prototype for check_inputs.c

11. In check_inputs.c, made logical changes to check combinations of transition
times and duty cycle

12. Made other changes to check_inputs.c to make error checking more robust and throrough