1. Removed files add_units.c and sort_array_by_column_number.c. add_units.c will be replaced by add_units2.c and function sort_array_by_column_number() is not used.
2. Updated check_executable.c to change use of sprintf() function to
use snprintf() function to avoid a memory overwrite.

3. Updated all function calls to add_units() in check_vpulse_inputs.c to
function calls of add_units_2(). The change adds ",0,0," to follow the number
of digits in the precision setting for the number conversion.

4. Updated all function calls to add_units() in main.c to
function calls of add_units_2(). The change adds ",0,0," to follow the number
of digits in the precision setting for the number conversion.

5. In main.c, moved dynamic memory allocation free() statements to end of file.
In prior location, the memory was overwritten after it was freed and was showing
a malloc error. After moving it to end of main.c, the freed memory will no longer be overwritten as the main program exits.

6. Updated makefile to remove source for sort_array_by_column_number.o and
add_units.o. Added source code for add_units2.o.

7. Updated remove_carriage_return.c to include use of newline
character '\r' as well as '\n'. This allows for correctly parsing
windows based files.

8. Updated .gitignore to include new example and Documentation directories

9. Updated README.txt file for v2.10 and to reflect added directories example and
Documentation

10. Updated vpulse.h as follows:
	a. Removed function prototype for add_units.c and added
	function prototype for add_units_2.c.
	b. Removed function prototype for sort_array_by_column_number.c
	as function is not used.
	c. Removed comment line above sort_array_by_column_number()
	function prototype "/*Analysis data related functions*/".
	d. Updated version and date from v2.09 3/24/2024 to v2.10 4/30/2024